Término da conexão com MSPedidos com sucesso e andamento da exclusão de itens no carrinho

+++ src/main/java/br/com/fiap/mscarrinho/domain/dto/ItemDtoRequest.java
@@ -2,6 +2,6 @@ package br.com.fiap.mscarrinho.domain.dto;
 
 public record ItemDtoRequest(
     Long idProduto,
-    int quantidade
+    Long quantidade
 ){}
